refactor: consolidate Tavily tools into single file
- merge Tavily search/extract/crawl/map and HTTP helper into tools/tavily.ts

- remove split Tavily tool files

- stop extending plan-mode metadata for Tavily sub-tools

- keep built-in tool registration via consolidated imports
