+++ packages/engine/src/tools/index.ts
@@ -4,10 +4,7 @@ import { EditTool } from './edit';
 import { GrepTool } from './grep';
 import { LsTool } from './ls';
 import { BashTool } from './bash';
-import { TavilyTool } from './tavily';
-import { TavilyExtractTool } from './tavily-extract';
-import { TavilyCrawlTool } from './tavily-crawl';
-import { TavilyMapTool } from './tavily-map';
+import { TavilyTool, TavilyExtractTool, TavilyCrawlTool, TavilyMapTool } from './tavily';
 import { GeminiProImageTool } from './gemini-pro-image';
 import { ClarifyTool } from './clarify';
 import { Tool } from 'ai';
